refactor(proto): move Boa runtime to `jstz_proto::runtime::v1`
