+++ crates/jstz_kernel/Cargo.toml
@@ -15,7 +15,6 @@ crate-type = ["cdylib", "lib"]
 
 [dependencies]
 bincode.workspace = true
-boa_engine.workspace = true
 http-serde.workspace = true
 http.workspace = true
 jstz_api = { path = "../jstz_api" }
