+++ crates/jstz_node/src/services/accounts.rs
@@ -5,11 +5,10 @@ use axum::{
 };
 use jstz_core::BinEncodable;
 use jstz_proto::{
-    api::KvValue,
     context::account::{
-        Account, Nonce, ParsedCode, SmartFunctionAccount, UserAccount,
-        ACCOUNTS_PATH_PREFIX,
+        Account, Nonce, SmartFunctionAccount, UserAccount, ACCOUNTS_PATH_PREFIX,
     },
+    runtime::{KvValue, ParsedCode},
 };
 use octez::OctezRollupClient;
 use serde::Deserialize;
