--- crates/jstz_proto/src/api/mod.rs
@@ -1,7 +1,0 @@
-mod kv;
-mod ledger;
-mod smart_function;
-
-pub use kv::{Kv, KvApi, KvValue};
-pub use ledger::LedgerApi;
-pub use smart_function::{SmartFunctionApi, TraceData};
