+++ crates/jstz_proto/src/context/account.rs
@@ -1,23 +1,24 @@
 use std::{
     fmt::{self, Display},
-    result,
     str::FromStr,
 };
 
-use crate::error::{Error, Result};
 use bincode::{Decode, Encode};
-use boa_engine::{Context, JsError, JsResult, Module, Source};
-use boa_gc::{empty_trace, Finalize, Trace};
 use jstz_core::{
     host::HostRuntime,
     kv::{Entry, Transaction},
 };
-use jstz_crypto::hash::Hash;
-use jstz_crypto::public_key_hash::PublicKeyHash;
-use jstz_crypto::smart_function_hash::SmartFunctionHash;
+use jstz_crypto::{
+    hash::Hash, public_key_hash::PublicKeyHash, smart_function_hash::SmartFunctionHash,
+};
 use serde::{Deserialize, Serialize};
 use tezos_smart_rollup::storage::path::{self, OwnedPath, RefPath};
 use utoipa::ToSchema;
+
+use crate::{
+    error::{Error, Result},
+    runtime::ParsedCode,
+};
 
 pub type Amount = u64;
 
@@ -51,36 +52,6 @@ impl Display for Nonce {
         self.0.fmt(f)
     }
 }
-
-// Invariant: if code is present it parses successfully
-#[derive(
-    Default, PartialEq, Eq, Debug, Clone, Serialize, Deserialize, ToSchema, Encode, Decode,
-)]
-#[schema(
-    format = "javascript",
-    example = "export default (request) => new Response('Hello world!')"
-)]
-pub struct ParsedCode(pub String);
-impl From<ParsedCode> for String {
-    fn from(ParsedCode(code): ParsedCode) -> Self {
-        code
-    }
-}
-impl Display for ParsedCode {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> result::Result<(), fmt::Error> {
-        Display::fmt(&self.0, formatter)
-    }
-}
-impl TryFrom<String> for ParsedCode {
-    type Error = JsError;
-    fn try_from(code: String) -> JsResult<Self> {
-        let src = Source::from_bytes(code.as_bytes());
-        let mut context = Context::default();
-        Module::parse(src, None, &mut context)?;
-        Ok(Self(code))
-    }
-}
-
 pub enum AddressKind {
     User,
     SmartFunction,
@@ -126,27 +97,13 @@ impl Addressable for PublicKeyHash {
 }
 
 #[derive(
-    Debug,
-    Clone,
-    PartialEq,
-    Eq,
-    Hash,
-    Serialize,
-    Deserialize,
-    Finalize,
-    ToSchema,
-    Encode,
-    Decode,
+    Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize, ToSchema, Encode, Decode,
 )]
 #[schema(description = "Tezos Address")]
 #[serde(untagged)]
 pub enum Address {
     User(PublicKeyHash),
     SmartFunction(SmartFunctionHash),
-}
-
-unsafe impl Trace for Address {
-    empty_trace!();
 }
 
 impl Display for Address {
@@ -320,11 +277,11 @@ impl Account {
         hrt: &impl HostRuntime,
         tx: &'a mut Transaction,
         addr: &SmartFunctionHash,
-    ) -> Result<&'a str> {
+    ) -> Result<&'a ParsedCode> {
         let account = Self::get_mut(hrt, tx, addr)?;
         match account {
             Self::SmartFunction(SmartFunctionAccount { function_code, .. }) => {
-                Ok(&function_code.0)
+                Ok(function_code)
             }
             Self::User(_) => Err(Error::AddressTypeMismatch),
         }
@@ -779,7 +736,7 @@ mod test {
 
             // Test empty initial code
             let code = Account::function_code(&host, &mut tx, sf_hash).unwrap();
-            assert_eq!(code, "");
+            assert_eq!(code.0, "");
 
             // Test empty function code
             assert!(
@@ -797,7 +754,7 @@ mod test {
             )
             .is_ok());
             let updated_code = Account::function_code(&host, &mut tx, sf_hash).unwrap();
-            assert_eq!(updated_code, valid_code);
+            assert_eq!(updated_code.0, valid_code);
 
             let account = Account::get_mut(&host, &mut tx, &user_addr).unwrap();
             assert!(matches!(account, Account::User(_)));
