+++ crates/jstz_proto/src/executor/smart_function/mod.rs
@@ -1,0 +1,5 @@
+pub mod deploy;
+pub mod host;
+pub mod run;
+
+pub use deploy::deploy_smart_function as deploy;
