+++ crates/jstz_proto/src/lib.rs
@@ -1,10 +1,8 @@
-pub mod api;
-mod error;
-
 pub mod context;
+mod error;
 pub mod executor;
-pub mod js_logger;
 pub mod operation;
 pub mod receipt;
 pub mod request_logger;
+pub mod runtime;
 pub use error::{Error, Result};
